Create Day 45 - ListManager
Your system should:
Have a menu that asks if you want to add, view, move or edit a 'to do'.
If you choose 'add' then the system should:

Prompt you to input what the to do is, when it is due by and the priority (high, medium or low).
Add the 'to do' to the list.
'View' should give two options:

View all - shows all 'to dos' with a pretty print.
View priority - allows you to search for high, medium or low priority and only see matching tasks.
'Edit' allows you to change any of the information within one of the 'to dos'.

'Remove' lets you completely remove a 'to do' when it is 'to done'.

Example:

🌟Life Organizer🌟
Welcome to the life organizer. Do you want to add, view, edit or remove a to do? > Add
What is the task? > Pay teachers more.
When is it due by? > 11/01/2022
What is the priority? >  High
Thanks, this task has been added.
Do you want to see the menu again or quit? > quit.

Font: https://replit.com/@prdmelo/Day45100Days